optimised CreatePage and used ContextAPi in Layout component
